JMdict: implement trie
When there isn't an exact result, return the longest result that is a
substring of the searched word and starts at the same point. (Same as
removing one character from the end of the query until it matches a word
but works differently)

When an exact result is found but there are longer words that start with
the query, return them as well.

Not used on the clientside yet.