add phrase lookup
When searching something from JMDict, there will be a button under the
results that, when clicked, will search up to 5 randomly chosen sentences
demonstrating the use of the word/phrase. If that's not enough, the results
can be browsed 20 at a time in the order they appear in wwwjdic.csv.

Particularly useful when a phrase is not found in JMDict.