radical input: filter results with jmdict
Characters aren't added one by one anymore. Instead, you can now input
one character or some of its radicals and then add another character to
the play with [+]. This immediately suggests all kanji that come after
the first (partial) character. The more info is supplied for either
character, the easier it is to find the correct character to any
position.

Example:
There is a word of 2 kanji. The first one has 亠 and the second has 十.
First, click 亠. Then, click [+]. Then, click 十. In the suggestions,
the wanted character, 革, appears fairly early (12.). You click it.
Then, you click [亠] (not in the radical list but above it). Now there
are only 3 characters left to choose from despite only supplying one
radical. You choose the wanted character, 変. Now that the word (変革)
is complete, you click [✓]. The word is added to the input field and
radical input is reset.