radical input: group by shape and stroke count
Results are divided in 4 columns by shape like similar kanji. In
addition, the kanji in each column are grouped by stroke count ranges
in an ascending order so that if a group has over 10 kanji in it, the
next stroke count will start a new group. The kanji in each group are
sorted and highlighted by frequency like before.